\name{Sim.DiffProc-package}
\alias{Sim.DiffProc-package}
\alias{Sim.DiffProc}
\docType{package}
\title{
Simulation of Diffusion Processes.
}
\description{
Simulation of diffusion processes and numerical solution of stochastic differential equations. Analysis of discrete-time approximations for stochastic differential equations (SDE) driven by Wiener processes,in financial and actuarial modeling and other areas of application.
}
\details{
\tabular{ll}{
Package: \tab Sim.DiffProc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-12-28\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
BOUKHETALA Kamal <kboukhetala@usthb.dz>, GUIDOUM Arsalane <starsalane@gmail.com>. 

Maintainer: GUIDOUM Arsalane <starsalane@gmail.com>
}
\references{
\enumerate{\item Franck Jedrzejewski. Modeles aleatoires et physique probabiliste, Springer, 2009.
\item T. Rolski, H. Schmidli, V. Schmidt and J. Teugels, Stochastic Processes for Insurance and Finance,John Wiley & Sons, 1998.
\item Fima C Klebaner. Introduction to stochastic calculus with application (Second Edition), Imperial College Press (ICP), 2005.
\item LAWRENCE C.EVANS. An introduction to stochastic differential equations (Version 1.2), Department of Mathematics (UC BERKELEY).
\item Hui-Hsiung Kuo. Introduction to stochastic integration, Springer, 2006.
\item E.Allen. Modeling with Ito stochastic differential equations, Springer, 2007.
\item Peter E.Kloeden, Eckhard Platen, Numerical solution of stochastic differential equations, Springer, 1995.
\item Douglas Henderson, Peter Plaschko, Stochastic differential equations in science and engineering,World Scientific, 2006.
\item A.Greiner, W.Strittmatter,and J.Honerkamp, Numerical Integration of Stochastic Differential Equations, Journal of Statistical Physics, Vol. 51, Nos. 1/2, 1988.
\item YOSHIHIRO SAITO, TAKETOMO MITSUI, SIMULATION OF STOCHASTIC DIFFERENTIAL EQUATIONS, Ann.Inst.Statist.Math, Vol. 45, No.3,419-432 (1993).
\item FRANCOIS-ERIC RACICOT, RAYMOND THEORET, Finance computationnelle et gestion des risques, Presses de universite du Quebec, 2006.
\item Avner Friedman, Stochastic differential equations and applications, Volume 1, ACADEMIC PRESS, 1975.
}
}
\keyword{ Diffusion Process }
\keyword{ Environment R }
\keyword{ Simulation }
\keyword{ financial models }
\keyword{ Stochastic Differential Equation }
\keyword{Numerical Solution of Stochastic Differential Equation}
\examples{
demo(BM2D)
demo(BMEuler)
demo(sim.sde)
demo(COVBM)
}
