\name{Sim.DiffProc-package}
\alias{Sim.DiffProc-package}
\alias{Sim.DiffProc}
\docType{package}
\title{
Simulation of Diffusion Processes.
}
\description{
The package Sim.DiffProc is an object created in the R language for simulation and modeling of stochastic differential equations (SDEs), 
and statistical analysis of diffusion processes solution of SDEs. This package contains many objects (code/function), for example a numerical 
methods to find the solutions to SDEs (one, two and three dimensional), which simulates a flows trajectories, with good accuracy. Many theoretical 
problems on the SDEs have become the object of practical research, as statistical analysis and simulation of solution of SDEs, enabled many 
searchers in different domains to use these equations to modeling and to analyse practical problems, in financial and actuarial modeling and 
other areas of application, for example modeling and simulate of dispersion in shallow water using the attractive center (Boukhetala K, 1996), 
and the stochastic calculus are applied to the random oscillators problem in physics. We hope that the package presented here and the updated 
survey on the subject might be of help for practitioners, postgraduate and PhD students, and researchers in the field who might want to implement 
new methods and ideas using R as a statistical environment. 
}
\details{
\tabular{ll}{
Package: \tab Sim.DiffProc\cr
Type: \tab Package\cr
Version: \tab 2.5\cr
Date: \tab 2012-06-05\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Boukhetala Kamal <kboukhetala@usthb.dz>, Guidoum Arsalane <starsalane@gmail.com>.

Maintainer: Arsalane Guidoum <starsalane@gmail.com>
}
\references{
\enumerate{
\item A. Greiner, W. Strittmatter,and J. Honerkamp, Numerical Integration of Stochastic Differential Equations, Journal of Statistical Physics, Vol. 51, Nos. 1/2, 1988.
\item A. Friedman, Stochastic differential equations and applications, Volume 1, ACADEMIC PRESS, 1975.
\item A C. GUIDOUM, Simulation des equations differentielles stochastiques sous R, theorie et applications. editions universitaires europeennes (French Edition). Allemagne, ISBN: 978-613-1-57992-9, 2012.
\item A C. GUIDOUM, Conception d un pro logiciel interactif sous R pour la simulation de processus de diffusion. Magister thesis, University of Science and Technology Houari Boumediene, N d ordre : 26/2012-M/MT, 2012.
\item D. Henderson, P. Plaschko, Stochastic differential equations in science and engineering,World Scientific, 2006.
\item E. Allen. Modeling with Ito stochastic differential equations, Springer, 2007.
\item F. Jedrzejewski. Modeles aleatoires et physique probabiliste, Springer, 2009.
\item F C Klebaner. Introduction to stochastic calculus with application (Second Edition), Imperial College Press (ICP), 2005.
\item Heinz S. Mathematical Modeling. Stochastic Evolution, pp. 295-334, Springer-Verlag, Berlin Heidelberg. ISBN 978-3-642-20310-7 2011.
\item Hui-Hsiung Kuo. Introduction to stochastic integration, Springer, 2006.
\item Ito K. Stochastic integral Tokyo. Proc. Jap. Acad, 20, pp. 519-529, 1944.
\item K. Boukhetala, Application des Processus de Diffusion, Echantillonnage Optimal. Ph.D. thesis, University of Science and Technology Houari Boumediene, BP 32 El-Alia, U.S.T.H.B, Algeria. 1998
\item K. Boukhetala, Estimation of the first passage time distribution for a simulated diffusion process, Maghreb Math.Rev, Vol.7, No 1, Jun 1998, pp. 1-25. 
\item K. Boukhetala, Simulation study of a dispersion about an attractive centre. In proceedings of 11th Symposium Computational Statistics, edited by R.Dutter and W.Grossman, Wien , Austria, 1994, pp. 128-130. 
\item K. Boukhetala, Modelling and simulation of a dispersion pollutant with attractive centre, Edited by Computational Mechanics Publications, Southampton ,U.K and Computational Mechanics Inc, Boston, USA, pp. 245-252. 
\item K. Boukhetala, Kernel density of the exit time in a simulated diffusion, les Annales Maghrebines De L ingenieur, Vol , 12, N Hors Serie. Novembre 1998, Tome II, pp 587-589. 
\item L C. Evans. An introduction to stochastic differential equations (Version 1.2), Department of Mathematics (UC BERKELEY).
\item Peter E. Kloeden, Eckhard Platen, Numerical solution of stochastic differential equations, Springer, 1995.
\item Racicot F E, Theoret R, Finance computationnelle et gestion des risques, Presses de universite du Quebec, 2006.
\item R Development Core Team (2012). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. ISBN 3-900051-07-0, URL http://www.R-project.org/.
\item Risken H. The Fokker Planck Equation : Methods of Solutions and Applications. 2nd edition, Springer Series in Synergetics. ISBN 9783540615309. 2001.
\item Saito Y, Mitsui T. Simulation of Stochastic Differential Equations. The Annals of the Institute of Statistical Mathematics, 3, pp. 419-432, 1993
\item Stefano M. Simulation and Inference for Stochastic Differential Equations, Example with R. Springer-Verlag, New York. ISBN 978-0-387-75838-1. 2008.
\item T. Rolski, H. Schmidli, V. Schmidt and J. Teugels, Stochastic Processes for Insurance and Finance,John Wiley & Sons, 1998.
}
}
\keyword{ Stochastic Differential Equation }
\keyword{ Diffusion Process Multidimensional}
\keyword{ Diffusion Process }
\keyword{ Actuarial Modeling }
\keyword{ Financial Models }
\keyword{ Stochastics Oscillators }
\keyword{ Numerical methods of SDE One-Dimensional}
\keyword{ Numerical methods of SDE Two-Dimensional}
\keyword{ Numerical methods of SDE Three-Dimensional}
\keyword{ Statistical Analysis }
\keyword{ Attractive Model}
\keyword{ Environment R }
\keyword{ Simulation }


